requirements.txt
Appium-Python-Client
pytest
behave
sauceclient
saucestorage
pyotp==2.2.6